Conflicts:
	web/src/main/java/org/fao/geonet/GeonetContext.java
	web/src/main/java/org/fao/geonet/Geonetwork.java
	web/src/main/java/org/fao/geonet/kernel/DataManager.java
	web/src/main/java/org/fao/geonet/kernel/SchemaManager.java
	web/src/main/java/org/fao/geonet/kernel/XmlSerializer.java
	web/src/main/java/org/fao/geonet/kernel/harvest/harvester/webdav/Harvester.java
	web/src/main/java/org/fao/geonet/kernel/mef/Importer.java
	web/src/main/java/org/fao/geonet/kernel/search/SearchManager.java
	web/src/main/java/org/fao/geonet/kernel/search/spatial/FullScanFilter.java
	web/src/main/java/org/fao/geonet/kernel/search/spatial/OgcGenericFilters.java
	web/src/main/java/org/fao/geonet/services/metadata/AjaxEditUtils.java
	web/src/main/java/org/fao/geonet/services/metadata/EditUtils.java
	web/src/main/java/org/fao/geonet/services/metadata/Insert.java
	web/src/main/java/org/fao/geonet/services/metadata/UpdateAdminOper.java
	web/src/main/java/org/fao/geonet/services/metadata/UpdateCategories.java
	web/src/main/java/org/fao/geonet/services/metadata/XslProcessing.java
	web/src/main/java/org/fao/geonet/services/ownership/Transfer.java
	web/src/main/java/org/fao/geonet/services/user/Remove.java
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-ar-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-cn-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-de-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-en-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-es-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-fi-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-fr-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-nl-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-no-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-pt-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-ru-default.sql
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/loc-vi-default.sql
	web/src/main/webapp/WEB-INF/config-csw.xml
	web/src/main/webapp/WEB-INF/server.prop

+++ web/src/main/java/org/fao/geonet/services/user/Remove.java
@@ -32,6 +32,8 @@ import jeeves.server.context.ServiceContext;
 import jeeves.utils.Util;
 
 import org.fao.geonet.constants.Geocat;
+import org.fao.geonet.GeonetContext;
+import org.fao.geonet.kernel.DataManager;
 import org.fao.geonet.constants.Geonet;
 import org.fao.geonet.constants.Params;
 import org.fao.geonet.kernel.reusable.ReusableTypes;
@@ -77,6 +79,9 @@ public class Remove implements Service
 		String      myProfile = usrSess.getProfile();
 		String      myUserId  = usrSess.getUserId();
 
+		GeonetContext gc = (GeonetContext) context.getHandlerContext(Geonet.CONTEXT_NAME);
+		DataManager dataMan = gc.getDataManager();
+
 		if (myUserId.equals(id)) {
 			throw new IllegalArgumentException("You cannot delete yourself from the user database");
 		}
@@ -91,6 +96,17 @@ public class Remove implements Service
 				if (adminlist.size() == 0) {
 				  throw new IllegalArgumentException("You don't have rights to delete this user because the user is not part of your group");
 				}
+			}
+
+			// Before processing DELETE check that the user is not referenced as an
+			// elsewhere in the GeoNetwork database - an exception is thrown if
+			// this is the case
+			if (dataMan.isUserMetadataOwner(dbms, new Integer(id))) {
+				throw new IllegalArgumentException("Cannot delete a user that is also a metadata owner");
+			}
+
+			if (dataMan.isUserMetadataStatus(dbms, new Integer(id))) {
+				throw new IllegalArgumentException("Cannot delete a user that has set a metadata status");
 			}
 
 			dbms.execute ("DELETE FROM UserGroups WHERE userId=" + id);
